+++ iextrading4j-acceptance/src/test/java/pl/zankowski/iextrading4j/test/acceptance/v1/SystemAcceptanceTest.java
@@ -12,6 +12,7 @@ public class SystemAcceptanceTest extends IEXCloudV1AcceptanceTestBase {
     public void systemStatusTest() {
         final SystemStatus systemStatus = cloudClient.executeRequest(new SystemStatusRequestBuilder()
                 .build());
+        System.out.println(systemStatus);
         assertThat(systemStatus).isNotNull();
     }