+++ iextrading4j-api/src/main/java/pl/zankowski/iextrading4j/api/system/SystemStatus.java
@@ -3,12 +3,13 @@ package pl.zankowski.iextrading4j.api.system;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.google.common.base.MoreObjects;
-import com.google.common.base.Objects;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+import java.util.Objects;
+import java.util.StringJoiner;
 
-@JsonPropertyOrder({"status", "version", "time"})
+@JsonPropertyOrder({"status", "version", "time", "currentMonthAPICalls"})
 public class SystemStatus implements Serializable {
 
     private static final long serialVersionUID = 3320852096256614294L;
@@ -16,15 +17,18 @@ public class SystemStatus implements Serializable {
     private final String status;
     private final String version;
     private final Long time;
+    private final BigDecimal currentMonthAPICalls;
 
     @JsonCreator
     public SystemStatus(
             @JsonProperty("status") final String status,
             @JsonProperty("version") final String version,
-            @JsonProperty("time") final Long time) {
+            @JsonProperty("time") final Long time,
+            @JsonProperty("currentMonthAPICalls") final BigDecimal currentMonthAPICalls) {
         this.status = status;
         this.version = version;
         this.time = time;
+        this.currentMonthAPICalls = currentMonthAPICalls;
     }
 
     public String getStatus() {
@@ -39,32 +43,33 @@ public class SystemStatus implements Serializable {
         return time;
     }
 
+    public BigDecimal getCurrentMonthAPICalls() {
+        return currentMonthAPICalls;
+    }
+
     @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final SystemStatus systemStatus1 = (SystemStatus) o;
-        return Objects.equal(status, systemStatus1.status) &&
-                Objects.equal(version, systemStatus1.version) &&
-                Objects.equal(time, systemStatus1.time);
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        SystemStatus that = (SystemStatus) o;
+        return Objects.equals(status, that.status) &&
+                Objects.equals(version, that.version) &&
+                Objects.equals(time, that.time) &&
+                Objects.equals(currentMonthAPICalls, that.currentMonthAPICalls);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(status, version, time);
+        return Objects.hash(status, version, time, currentMonthAPICalls);
     }
 
     @Override
     public String toString() {
-        return MoreObjects.toStringHelper(this)
-                .add("status", status)
-                .add("version", version)
-                .add("time", time)
+        return new StringJoiner(", ", SystemStatus.class.getSimpleName() + "[", "]")
+                .add("status='" + status + "'")
+                .add("version='" + version + "'")
+                .add("time=" + time)
+                .add("currentMonthAPICalls=" + currentMonthAPICalls)
                 .toString();
     }
-
 }