+++ iextrading4j-api/src/test/java/pl/zankowski/iextrading4j/api/system/SystemStatusTest.java
@@ -5,6 +5,8 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.Test;
 import pl.zankowski.iextrading4j.api.util.ToStringVerifier;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class SystemStatusTest {
@@ -16,12 +18,14 @@ public class SystemStatusTest {
         final String status = fixture.create(String.class);
         final String version = fixture.create(String.class);
         final Long time = fixture.create(Long.class);
+        final BigDecimal currentMonthAPICalls = fixture.create(BigDecimal.class);
 
-        final SystemStatus systemStatus = new SystemStatus(status, version, time);
+        final SystemStatus systemStatus = new SystemStatus(status, version, time, currentMonthAPICalls);
 
         assertThat(systemStatus.getStatus()).isEqualTo(status);
         assertThat(systemStatus.getVersion()).isEqualTo(version);
         assertThat(systemStatus.getTime()).isEqualTo(time);
+        assertThat(systemStatus.getCurrentMonthAPICalls()).isEqualTo(systemStatus.getCurrentMonthAPICalls());
     }
 
     @Test